regroup code that needs TILEDB_DEPRECATED
